created new file for help widget
